Enhance part and section management: update timestamps on create, update, and delete operations; improve error handling for related entities; increment total submissions count in submission creation.

+++ src/controllers/section.controller.js
@@ -1,5 +1,6 @@
 // Models
 const Part = require("../models/Part");
+const Test = require("../models/Test");
 const Section = require("../models/Section");
 
 // Helpers
@@ -11,6 +12,24 @@ const createSection = async (req, res, next) => {
   const { title, description, type, partId, testId, module } = req.body;
 
   try {
+    // Test
+    const test = await Test.findOne({ _id: testId, createdBy });
+    if (!test) {
+      return res.status(404).json({
+        code: "testNotFound",
+        message: "Test topilmadi",
+      });
+    }
+
+    // Part
+    const part = await Part.findOne({ _id: partId, createdBy });
+    if (!part) {
+      return res.status(404).json({
+        code: "partNotFound",
+        message: "Qism topilmadi",
+      });
+    }
+
     const section = await Section.create({
       type,
       title,
@@ -29,6 +48,9 @@ const createSection = async (req, res, next) => {
       code: "sectionCreated",
       message: "New section created successfully",
     });
+
+    test[module].updatedAt = Date.now();
+    await test.save();
   } catch (err) {
     next(err);
   }
@@ -92,23 +114,37 @@ const updateSection = async (req, res, next) => {
   const sectionData = pickAllowedFields(req.body, allowedFields);
 
   try {
-    const section = await Section.findOneAndUpdate(
-      { _id: id, createdBy },
-      sectionData,
-      { new: true }
-    );
-
+    // Section
+    const section = await Section.findOne({ _id: id, createdBy });
     if (!section) {
-      return res
-        .status(404)
-        .json({ code: "sectionNotFound", message: "Section not found" });
+      return res.status(404).json({
+        code: "sectionNotFound",
+        message: "Bo'lim topilmadi",
+      });
     }
 
-    res.json({
-      section,
-      code: "sectionUpdated",
-      message: "Section updated successfully",
+    // Test
+    const test = await Test.findOne({ _id: section.testId, createdBy });
+    if (!test) {
+      return res.status(404).json({
+        code: "testNotFound",
+        message: "Test topilmadi",
+      });
+    }
+
+    const updatedSection = await Section.findByIdAndUpdate(id, sectionData, {
+      new: true,
     });
+
+    res.json({
+      code: "sectionUpdated",
+      section: updatedSection,
+      message: "Bo'lim yangilandi",
+    });
+
+    // Update test's module updatedAt
+    test[section.module].updatedAt = Date.now();
+    await test.save();
   } catch (err) {
     next(err);
   }
@@ -120,6 +156,37 @@ const deleteSection = async (req, res, next) => {
   const createdBy = req.user._id;
 
   try {
+    // Section
+    const section = await Section.findOne({ _id: id, createdBy });
+    if (!section) {
+      return res.status(404).json({
+        code: "sectionNotFound",
+        message: "Bo'lim topilmadi",
+      });
+    }
+
+    // Test
+    const test = await Test.findOne({ _id: section.testId, createdBy });
+    if (!test) {
+      return res.status(404).json({
+        code: "testNotFound",
+        message: "Test topilmadi",
+      });
+    }
+
+    // Part
+    const part = await Part.findOneAndUpdate(
+      { _id: section.partId, createdBy },
+      { $pull: { sections: section._id } }
+    );
+    if (!part) {
+      return res.status(404).json({
+        code: "partNotFound",
+        message: "Qism topilmadi",
+      });
+    }
+
+    // Delete
     const deleted = await Section.findOneAndDelete({ _id: id, createdBy });
     if (!deleted) {
       return res.status(404).json({
@@ -129,9 +196,14 @@ const deleteSection = async (req, res, next) => {
     }
 
     res.json({
+      section: deleted,
       code: "sectionDeleted",
       message: "Bo'lim o'chirildi",
     });
+
+    // Update test's module updatedAt
+    test[section.module].updatedAt = Date.now();
+    await test.save();
   } catch (err) {
     next(err);
   }
